Move test to new file
